+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 ### Added
 
-- Lint check that ensures all function arguments are documented
+- Lint check that ensures all function arguments are documented (including
+  `*args` and `**kwargs`)
+- Support for skipping test functions in test files
 
 [//]: # "Release links"
 [v1.0.0]: https://github.com/jdkandersson/flake8-docstrings-complete/releases/v1.0.0